feat: Further component refactoring and hook extraction
